requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
cytoolz==0.11.2
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.0
djangorestframework-api-key==2.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.19.1
psycopg2==2.9.2
pycryptodome==3.12.0
pyrsistent==0.18.0
pytz==2021.3
pywin32==303
requests==2.26.0
rlp==2.0.1
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
varint==1.0.2
web3==5.25.0
websockets==9.1
whitenoise==5.3.0
yarl==1.7.2
